refactor(blocks): change hardness from Int to Float
Allows sub-integer break times (e.g. 0.5s leaves) without integer
quantization artifacts. Sentinel -1 (unbreakable) mapped to
Float.MAX_VALUE instead of Int.MAX_VALUE. BlockBreaker comparisons
and all protocol messages updated accordingly.

+++ app/webApp/src/wasmJsMain/kotlin/org/micoli/micraft/LocalPlayerController.kt
@@ -106,7 +106,7 @@ class LocalPlayerController(
     }
 
     private fun Double.r3(): String {
-        val v = (kotlin.math.round(this * 1000).toDouble() / 1000.0).toString()
+        val v = (kotlin.math.round(this * 1000) / 1000.0).toString()
         val dot = v.indexOf('.')
         return if (dot < 0) "$v.000" else v.padEnd(dot + 4, '0').take(dot + 4)
     }
